Refactor the SnackbarNotificationComponent: - rename into NoticeComponent; - require the string of the banner to be dismissed; - make more generic so that it is applicable to both permanent and dismissable banners.
