+++ packages/calypso-e2e/src/lib/components/index.ts
@@ -12,7 +12,7 @@ export * from './domain-search-component';
 export * from './isolated-block-editor-component';
 export * from './block-widget-editor-component';
 export * from './revisions-component';
-export * from './snackbar-notification-component';
+export * from './notice-component';
 export * from './page-template-modal-component';
 export * from './react-modal-component';
 export * from './editor-inline-block-inserter-component';
